requirements.txt
# Newsletter Automation Dependencies

# RSS Feed Parsing
feedparser==6.0.11

# HTTP Requests
requests==2.31.0

# Email Sending
resend==0.8.0

# LLM Integration via OpenRouter
openai==1.12.0

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0

# Article Content Extraction
readability-lxml==0.8.1
newspaper3k==0.2.8

# Date/Time handling
python-dateutil==2.8.2

# HTML Email Templates
jinja2==3.1.3

# Environment Variables
python-dotenv==1.0.1
